+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project aims to adhere to [Semantic Versioning](http://semver.org/).
 ## [3.2.2-SNAPSHOT] - Coming soon!
 ### Fixed
  - [Misleading log message is shown for retrying requests that are unretryable](https://github.com/joyent/java-manta/issues/388)
+ - [`MantaClientHttpResponseException` provide to way to get the response headers](https://github.com/joyent/java-manta/issues/383)
 
 ## [3.2.1] - 2017-12-01
 ### Added